Consolidate widget tests
Now that we've ported all the widget tests to fn3, we don't need a separate fn3
directory.
