+++ sky/unit/test/widget/block_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
 import '../engine/mock_events.dart';
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 final Key blockKey = new Key('test');
