+++ sky/unit/test/widget/box_decoration_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/material.dart';
 import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Circles can have uniform borders', () {
