+++ sky/unit/test/widget/center_test.dart
@@ -1,7 +1,7 @@
 import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Can be placed in an infinte box', () {
