+++ sky/unit/test/widget/gesture_detector_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
 import '../engine/mock_events.dart';
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('Uncontested scrolls start immediately', () {
