+++ sky/unit/test/widget/navigator_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/animation.dart';
 import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 class FirstComponent extends StatelessComponent {
   FirstComponent(this.navigator);
