+++ sky/unit/test/widget/progress_indicator_test.dart
@@ -3,7 +3,7 @@ import 'package:sky/rendering.dart';
 import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 void main() {
   test('LinearProgressIndicator changes when its value changes', () {
