+++ sky/unit/test/widget/set_state_test.dart
@@ -2,7 +2,7 @@ import 'package:sky/src/fn3.dart';
 import 'package:test/test.dart';
 
 import '../engine/mock_events.dart';
-import '../fn3/widget_tester.dart';
+import 'widget_tester.dart';
 
 class Inside extends StatefulComponent {
   InsideState createState() => new InsideState();
